Fix bug PrepareSample for action gesture recognizer
Don't know why, but initializing the list with capacity (N, for example) does not make the list filled with N null values, and Debug.Log said the list has 0 value.
I brought that function to the initialization method, so it will be initialized in Start(), and no need to reload it again when recognizing gestures.

Tested OK. Action can be started. Stopping action is not implemented.

Work to do: Stopping action.