Fix error that object following box disappear when gesture no longer stay.
Add another gameObject called "moving object", that will be the box's parent. So the BoxAnnotation will move the parent/wrapper object, while activating/disabling the box only. When the box touch any object that can be dragged along, it will be the child of the wrapper object, not the box. So we can disable the box without affecting the dragging object - after we disable the box, OnTriggerExit ensured -> the dragged object returned back to its original parent.