Fix: HandGestureActionRecognizer triggering action events in a different thread (due to being called by other threaded MyHandTrackingSolution)
So the Unity APIs can't be used if the method is being run in different threads other than main thread. Because of that, when action recognizer call 'TurnOffTheActions' for the first time (in the beginning, I didn't do any hand gesture), it leads to HandGestureInputEvent invoking action recognized event, which is a type of UnityEvent (from Unity APIs)! So it throw 'null reference' because in the other thread there is no Unity APIs so I guess it thought this type did not exist.

I changed this by implementing an Update() with SyncProcess() in it, inside sync process the actual gesture recognition takes place. It's similar to AnnotationControllers, where the DrawLater functions are called from the other thread by a solution file, then there's a loop in the main thread (LateUpdate) that actually control and render the annotations based on data received.