Make the game hand annotation be disabled when no longer detecting hand.
As far as I can understand, it's because of the currentObjectActiveness being sent all the time (every frame), and because its parameters will always be a List object, that means it never has null parameter in terms of activeness values. Hence, the function ActivateFor will always be called, even after the hands go out of screen.

Simply don't automatically activate if the activeness is set. That way, only the rest parameters can activate the annotation.