Update MagnetCube + magnetic to support destroying the magnetic object
Because the towel will be destroyed when the game is won, I must implement a few more lines to handle the case when a magnetic object is being destroyed: it will call the magnet cube to detach and nullify the variables. And that's all.

Current issues: as of this moment, and I don't know why, but in the third level the dirt cubes are not spawing correctly : three of them become empty concrete with no cube in it, and one giant cube with a scale of somewhat 11.6 ? I don't get it. Investigating.